my-agent/marketplace-specialist/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-generativeai==0.8.4
jsonschema==4.23.0
python-docx==1.1.2